Added a timer for the BAC update method. The timer executes, but we need to change the logic.
I created a 'drink' class which will house the data of what the user is drinking. It holds the time drank, Volume of drink, Name of drink, and BAC of the drink.

In main I created a list of drinks, this will be used to keep track of what the user has been drinking and be used to calculate the BAC.

The math will be derived from my private iOS version of this application. Simply, it's based on the time of the first drink, users' weight, height, and sex. Then all the BAC of drinks between then.

Also added an enforcement to the portrait mode of this app. User cannot go landscape mode for now.

Needs to be done:
 Settings (holds users weight, age, height, sex)
 Passing back drinks from change drink page and home.
 calculations
Extra
 List of drinks for future review by user. I.E. drink log